+++ libp2p/src/builder/phase/bandwidth_logging.rs
@@ -1,4 +1,5 @@
 use super::*;
+#[allow(deprecated)]
 use crate::bandwidth::BandwidthSinks;
 use crate::transport_ext::TransportExt;
 use crate::SwarmBuilder;
@@ -13,16 +14,18 @@ pub struct BandwidthLoggingPhase<T, R> {
 impl<T: AuthenticatedMultiplexedTransport, Provider, R>
     SwarmBuilder<Provider, BandwidthLoggingPhase<T, R>>
 {
+    #[allow(deprecated)]
+    #[deprecated(note = "Use `with_bandwidth_metrics` instead.")]
     pub fn with_bandwidth_logging(
         self,
     ) -> (
-        SwarmBuilder<Provider, BehaviourPhase<impl AuthenticatedMultiplexedTransport, R>>,
+        SwarmBuilder<Provider, BandwidthMetricsPhase<impl AuthenticatedMultiplexedTransport, R>>,
         Arc<BandwidthSinks>,
     ) {
         let (transport, sinks) = self.phase.transport.with_bandwidth_logging();
         (
             SwarmBuilder {
-                phase: BehaviourPhase {
+                phase: BandwidthMetricsPhase {
                     relay_behaviour: self.phase.relay_behaviour,
                     transport,
                 },
@@ -33,9 +36,9 @@ impl<T: AuthenticatedMultiplexedTransport, Provider, R>
         )
     }
 
-    pub fn without_bandwidth_logging(self) -> SwarmBuilder<Provider, BehaviourPhase<T, R>> {
+    pub fn without_bandwidth_logging(self) -> SwarmBuilder<Provider, BandwidthMetricsPhase<T, R>> {
         SwarmBuilder {
-            phase: BehaviourPhase {
+            phase: BandwidthMetricsPhase {
                 relay_behaviour: self.phase.relay_behaviour,
                 transport: self.phase.transport,
             },
@@ -46,6 +49,18 @@ impl<T: AuthenticatedMultiplexedTransport, Provider, R>
 }
 
 // Shortcuts
+#[cfg(feature = "metrics")]
+impl<Provider, T: AuthenticatedMultiplexedTransport, R>
+    SwarmBuilder<Provider, BandwidthLoggingPhase<T, R>>
+{
+    pub fn with_bandwidth_metrics(
+        self,
+        registry: &mut libp2p_metrics::Registry,
+    ) -> SwarmBuilder<Provider, BehaviourPhase<impl AuthenticatedMultiplexedTransport, R>> {
+        self.without_bandwidth_logging()
+            .with_bandwidth_metrics(registry)
+    }
+}
 #[cfg(feature = "relay")]
 impl<Provider, T: AuthenticatedMultiplexedTransport>
     SwarmBuilder<Provider, BandwidthLoggingPhase<T, libp2p_relay::client::Behaviour>>
@@ -54,10 +69,11 @@ impl<Provider, T: AuthenticatedMultiplexedTransport>
         self,
         constructor: impl FnOnce(&libp2p_identity::Keypair, libp2p_relay::client::Behaviour) -> R,
     ) -> Result<SwarmBuilder<Provider, SwarmPhase<T, B>>, R::Error> {
-        self.without_bandwidth_logging().with_behaviour(constructor)
+        self.without_bandwidth_logging()
+            .without_bandwidth_metrics()
+            .with_behaviour(constructor)
     }
 }
-
 impl<Provider, T: AuthenticatedMultiplexedTransport>
     SwarmBuilder<Provider, BandwidthLoggingPhase<T, NoRelayBehaviour>>
 {
@@ -65,6 +81,8 @@ impl<Provider, T: AuthenticatedMultiplexedTransport>
         self,
         constructor: impl FnOnce(&libp2p_identity::Keypair) -> R,
     ) -> Result<SwarmBuilder<Provider, SwarmPhase<T, B>>, R::Error> {
-        self.without_bandwidth_logging().with_behaviour(constructor)
+        self.without_bandwidth_logging()
+            .without_bandwidth_metrics()
+            .with_behaviour(constructor)
     }
 }